requirements.txt
flask
flask_limiter
pillow
qrcode